Pin MappaMundi commit (#380)
The Cartfile we inherited from Firefox contains this line:

`github "mozilla-mobile/MappaMundi"          "master"`

recently, MappaMundi was [updated to use the A-Star project](https://github.com/mozilla-mobile/MappaMundi/pull/17) for pathfinding. This project, however, needs Swift 5. Hence, the build is broken.

To fix, pin the MappaMundi version to a commit before they introduce the A-Star algorithm.